Fix timestamps on events
The event was reporting some crazy numbers from before 1970 and
sometimes with a 10 year difference between events that took place
within seconds.

The code was losing information as it was forcing a 64 bit number into a
32 bit number. This fix provides the correct timestamp in
nanoseconds.

+++ src/gpio_chip.h
@@ -5,6 +5,7 @@
 #define GPIO_CHIP_H
 
 #include <linux/gpio.h>
+#include <stdint.h>
 #include "erl_nif.h"
 
 #define GPIO_CHIP_LINE_INPUT 0x00
@@ -44,7 +45,7 @@ struct gpio_chip_event_handle {
  * Struct for event data
  */
 struct gpio_chip_event_data {
-    int timestamp;
+    uint64_t timestamp;
     int id;
 };
